fix: 오류 및 충돌 해결 (#5)
- pom.xml : <path> 오류
- CommentController, Service : CustomUserDetails 경로 수정
- EmotionStatRepository 추가
- DiaryMember : Builder 추가
- TagController : 태그제거 부분 entryId, tagId -> Long으로 변경

+++ backend/src/main/java/com/diary/domain/tag/controller/TagController.java
@@ -57,8 +57,8 @@ public class TagController {
     /** [4] 일기-태그 제거 **/
     @DeleteMapping("/entry/{entryId}/tags/{tagId}")
     public ResponseEntity<Map<String, String>> removeTagFromEntry(
-            @PathVariable Integer entryId,
-            @PathVariable Integer tagId
+            @PathVariable Long entryId,
+            @PathVariable Long tagId
     ) {
         // 일기에서 태그 제거 API
         tagService.removeTagFromEntry(entryId, tagId);
